extracted LambdaToFunc from OnnxRuntime for use also in OnnxPartialEvaluator
