+++ cr-examples/onnx/src/main/java/oracle/code/onnx/OnnxRuntime.java
@@ -4,22 +4,15 @@ import java.io.File;
 import java.io.IOException;
 import java.lang.foreign.*;
 import java.lang.invoke.MethodHandles;
-import java.lang.reflect.Method;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
 import java.util.*;
-import java.util.function.Predicate;
 import java.util.function.Supplier;
 import java.util.stream.IntStream;
 import jdk.incubator.code.*;
 
 import jdk.incubator.code.op.CoreOp;
-import jdk.incubator.code.type.ClassType;
-import jdk.incubator.code.type.FunctionType;
-import jdk.incubator.code.type.MethodRef;
-import jdk.incubator.code.type.VarType;
-import oracle.code.onnx.compiler.OnnxTransformer;
 import oracle.code.onnx.foreign.OrtApi;
 import oracle.code.onnx.foreign.OrtApiBase;
 
@@ -76,58 +69,10 @@ public final class OnnxRuntime {
 
         @Override
         protected CachedSession computeValue(Class<?> type) {
-            var lambda = (CoreOp.LambdaOp) q.op();
-
-            CoreOp.FuncOp onnxFunc;
-            int[] operandsMapping;
-
-            // Shortcut for lambda expressions that call just one method
-            if (singleMethodInvocation(lambda) instanceof
-                    SingleMethod(CoreOp.InvokeOp iop, Map<Value, Value> valueMapping)) {
-                Method m;
-                try {
-                    m = iop.invokeDescriptor().resolveToMethod(l, iop.invokeKind());
-                } catch (ReflectiveOperationException e) {
-                    throw new RuntimeException(e);
-                }
-
-                CoreOp.FuncOp f = Op.ofMethod(m).orElseThrow();
-                onnxFunc = OnnxTransformer.transform(l, f);
-
-                var operands = iop.operands();
-                var captured = q.capturedValues().sequencedKeySet().stream().toList();
-                operandsMapping = new int[iop.operands().size()];
-                for (int i = 0; i < operandsMapping.length; i++) {
-                    operandsMapping[i] = captured.indexOf(valueMapping.get(operands.get(i)));
-                }
-
-            } else {
-                var capturedValues = lambda.capturedValues();
-                var functionType = FunctionType.functionType(lambda.invokableType().returnType(),
-                        capturedValues.stream().map(Value::type)
-                                .map(t -> t instanceof VarType vt ? vt.valueType() : t).toList());
-                onnxFunc = OnnxTransformer.transform(l, CoreOp.func("onnxCode", functionType)
-                        .body(bb -> {
-                            bb.context().mapValues(capturedValues, bb.parameters());
-                            for (Op op : lambda.body().entryBlock().ops()) {
-                                int i;
-                                if (op instanceof CoreOp.VarAccessOp.VarLoadOp load &&
-                                        (i = capturedValues.indexOf(load.varOp().result())) >= 0) {
-                                    bb.context().mapValue(op.result(), bb.parameters().get(i)); // remap var load result to block param
-                                } else {
-                                    bb.apply(op);
-                                }
-                            }
-                        }));
-
-                operandsMapping = new int[capturedValues.size()];
-                for (int i = 0; i < operandsMapping.length; i++) {
-                    operandsMapping[i] = i;
-                }
-            }
+            var mf = LambdaToFunc.fromLambda(l, (CoreOp.LambdaOp)q.op());
             return new CachedSession(getInstance().createSession(
                     Arena.ofAuto(), // cached session must be created under its own auto arena
-                    OnnxProtoBuilder.build(onnxFunc.body().entryBlock())), operandsMapping);
+                    OnnxProtoBuilder.build(mf.func().body().entryBlock())), mf.operandsMapping());
         }
     }
 
@@ -150,97 +95,6 @@ public final class OnnxRuntime {
                 .toList();
 
         return model.session.run(arena, arguments).getFirst();
-    }
-
-    record SingleMethod(CoreOp.InvokeOp iop, Map<Value, Value> valueMapping) {}
-    static SingleMethod singleMethodInvocation(CoreOp.LambdaOp lop) {
-        // Single block
-        if (lop.body().blocks().size() > 1) {
-            return null;
-        }
-
-        Map<Value, Value> valueMapping = new HashMap<>();
-        CoreOp.InvokeOp methodRefInvokeOp = extractMethodInvoke(valueMapping, lop.body().entryBlock().ops());
-        if (methodRefInvokeOp == null) {
-            return null;
-        }
-
-        return new SingleMethod(methodRefInvokeOp, valueMapping);
-    }
-
-    static CoreOp.InvokeOp extractMethodInvoke(Map<Value, Value> valueMapping, List<Op> ops) {
-        CoreOp.InvokeOp methodRefInvokeOp = null;
-        for (Op op : ops) {
-            switch (op) {
-                case CoreOp.VarOp varOp -> {
-                    if (isValueUsedWithOp(varOp.result(), o -> o instanceof CoreOp.VarAccessOp.VarStoreOp)) {
-                        return null;
-                    }
-                }
-                case CoreOp.VarAccessOp.VarLoadOp varLoadOp -> {
-                    Value v = varLoadOp.varOp().result();
-                    valueMapping.put(varLoadOp.result(), valueMapping.getOrDefault(v, v));
-                }
-                case CoreOp.InvokeOp iop when isBoxOrUnboxInvocation(iop) -> {
-                    Value v = iop.operands().getFirst();
-                    valueMapping.put(iop.result(), valueMapping.getOrDefault(v, v));
-                }
-                case CoreOp.InvokeOp iop -> {
-                    if (methodRefInvokeOp != null) {
-                        return null;
-                    }
-
-                    for (Value o : iop.operands()) {
-                        valueMapping.put(o, valueMapping.getOrDefault(o, o));
-                    }
-                    methodRefInvokeOp = iop;
-                }
-                case CoreOp.ReturnOp rop -> {
-                    if (methodRefInvokeOp == null) {
-                        return null;
-                    }
-                    Value r = rop.returnValue();
-                    if (!(valueMapping.getOrDefault(r, r) instanceof Op.Result invokeResult)) {
-                        return null;
-                    }
-                    if (invokeResult.op() != methodRefInvokeOp) {
-                        return null;
-                    }
-                    assert methodRefInvokeOp.result().uses().size() == 1;
-                }
-                default -> {
-                    return null;
-                }
-            }
-        }
-
-        return methodRefInvokeOp;
-    }
-
-    private static boolean isValueUsedWithOp(Value value, Predicate<Op> opPredicate) {
-        for (Op.Result user : value.uses()) {
-            if (opPredicate.test(user.op())) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    // @@@ Move to functionality on JavaType(s)
-    static final Set<String> UNBOX_NAMES = Set.of(
-            "byteValue",
-            "shortValue",
-            "charValue",
-            "intValue",
-            "longValue",
-            "floatValue",
-            "doubleValue",
-            "booleanValue");
-
-    private static boolean isBoxOrUnboxInvocation(CoreOp.InvokeOp iop) {
-        MethodRef mr = iop.invokeDescriptor();
-        return mr.refType() instanceof ClassType ct && ct.unbox().isPresent() &&
-                (UNBOX_NAMES.contains(mr.name()) || mr.name().equals("valueOf"));
     }
 
     public static OnnxRuntime getInstance() {
