+++ cr-examples/onnx/src/main/java/oracle/code/onnx/compiler/OnnxPartialEvaluator.java
@@ -37,6 +37,7 @@ import java.lang.reflect.Array;
 import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import oracle.code.onnx.LambdaToFunc;
 
 final class OnnxPartialEvaluator {
 
@@ -474,25 +475,8 @@ final class OnnxPartialEvaluator {
                         .collect(Collectors.joining());
             }
             case CoreOp.LambdaOp lambdaOp -> {
-                // @@@ share with runtime
                 // @@@ handle captured values
-                var capturedValues = lambdaOp.capturedValues();
-                var functionType = FunctionType.functionType(lambdaOp.invokableType().returnType(),
-                        capturedValues.stream().map(Value::type)
-                                .map(t -> t instanceof VarType vt ? vt.valueType() : t).toList());
-                return OnnxTransformer.transform(l, CoreOp.func("onnxCode", functionType)
-                        .body(bb -> {
-                            bb.context().mapValues(capturedValues, bb.parameters());
-                            for (Op op : lambdaOp.body().entryBlock().ops()) {
-                                int i;
-                                if (op instanceof CoreOp.VarAccessOp.VarLoadOp load &&
-                                        (i = capturedValues.indexOf(load.varOp().result())) >= 0) {
-                                    bb.context().mapValue(op.result(), bb.parameters().get(i)); // remap var load result to block param
-                                } else {
-                                    bb.apply(op);
-                                }
-                            }
-                        })).body().copy(CopyContext.create());
+                return LambdaToFunc.fromLambda(l, lambdaOp).func().body().copy(CopyContext.create());
             }
             case null, default -> throw interpreterException(
                     new UnsupportedOperationException("Unsupported operation: " + o.opName()));
